use trait
